fixes reducer hot reload
